Move Action, Category and CoreCommand to cli module
