+++ src/_repobee/cli/dispatch.py
@@ -38,10 +38,10 @@ def dispatch_command(
     """
     hook_results = {}
     dispatch_table = {
-        plug.CoreCommand.repos: _dispatch_repos_command,
-        plug.CoreCommand.issues: _dispatch_issues_command,
-        plug.CoreCommand.config: _dispatch_config_command,
-        plug.CoreCommand.reviews: _dispatch_reviews_command,
+        plug.cli.CoreCommand.repos: _dispatch_repos_command,
+        plug.cli.CoreCommand.issues: _dispatch_issues_command,
+        plug.cli.CoreCommand.config: _dispatch_config_command,
+        plug.cli.CoreCommand.reviews: _dispatch_reviews_command,
     }
 
     is_ext_command = "_extension_command" in args
@@ -50,15 +50,15 @@ def dispatch_command(
         res = ext_cmd.callback(args, api)
         hook_results = {ext_cmd.name: [res]} if res else hook_results
     else:
-        category = plug.CoreCommand(args.category)
+        category = plug.cli.CoreCommand(args.category)
         hook_results = (
             dispatch_table[category](args, config_file, api) or hook_results
         )
 
     if is_ext_command or args.action in [
-        plug.CoreCommand.repos.setup.name,
-        plug.CoreCommand.repos.update.name,
-        plug.CoreCommand.repos.clone.name,
+        plug.cli.CoreCommand.repos.setup.name,
+        plug.cli.CoreCommand.repos.update.name,
+        plug.cli.CoreCommand.repos.clone.name,
     ]:
         LOGGER.info(formatters.format_hook_results_output(hook_results))
     if hook_results and "hook_results_file" in args and args.hook_results_file:
@@ -72,7 +72,7 @@ def dispatch_command(
 def _dispatch_repos_command(
     args: argparse.Namespace, config_file: pathlib.Path, api: plug.API
 ) -> Optional[Mapping[str, List[plug.Result]]]:
-    repos = plug.CoreCommand.repos
+    repos = plug.cli.CoreCommand.repos
     action = repos[args.action]
     if action == repos.setup:
         return command.setup_student_repos(
@@ -97,7 +97,7 @@ def _dispatch_repos_command(
 def _dispatch_issues_command(
     args: argparse.Namespace, config_file: pathlib.Path, api: plug.API
 ) -> Optional[Mapping[str, List[plug.Result]]]:
-    issues = plug.CoreCommand.issues
+    issues = plug.cli.CoreCommand.issues
     action = issues[args.action]
     if action == issues.open:
         command.open_issue(
@@ -122,7 +122,7 @@ def _dispatch_issues_command(
 def _dispatch_config_command(
     args: argparse.Namespace, config_file: pathlib.Path, api: plug.API
 ) -> Optional[Mapping[str, List[plug.Result]]]:
-    config = plug.CoreCommand.config
+    config = plug.cli.CoreCommand.config
     action = config[args.action]
     if action == config.verify:
         plug.manager.hook.get_api_class().verify_settings(
@@ -142,7 +142,7 @@ def _dispatch_config_command(
 def _dispatch_reviews_command(
     args: argparse.Namespace, config_file: pathlib.Path, api: plug.API
 ) -> Optional[Mapping[str, List[plug.Result]]]:
-    reviews = plug.CoreCommand.reviews
+    reviews = plug.cli.CoreCommand.reviews
     action = reviews[args.action]
     if action == reviews.assign:
         command.assign_peer_reviews(
